2010-11-09  Jim Huang  <jserv@0xlab.org>
        * src/common/key2pho.c,
        src/common/chewing-utf8-util.c,
        src/hanyupinyin.c,
        src/char.c,
        src/tools/sort_dic.c,
        src/chewingutil.c,
        src/zuin.c: Fix compilation warning with gcc-4.5


git-svn-id: https://svn.csie.net/chewing/libchewing/trunk@1080 ac1be623-90ea-0310-9410-ba68b2efa777

+++ src/char.c
@@ -30,7 +30,7 @@
 #ifdef USE_BINARY_DATA
 static uint16* arrPhone = NULL;
 static int *begin = NULL;
-static int phone_num;
+static size_t phone_num;
 static plat_mmap char_begin_mmap;
 static plat_mmap char_phone_mmap;
 static void *dict = NULL;
